feat: publish xstyled utilities as a separated package
