+++ packages/core/src/index.js
@@ -1,4 +1,3 @@
 export * from './createBox'
 export * from './propGetters'
 export * from './transform'
-export * from './util'
