+++ packages/styled-components/src/css.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-continue, no-loop-func, no-cond-assign */
 import { css as scCss } from 'styled-components'
-import { transform, flattenStrings } from '@xstyled/core'
+import { flattenStrings } from '@xstyled/util'
+import { transform } from '@xstyled/core'
 
 export function css(...rawArgs) {
   const scCssArgs = scCss(...rawArgs)
