+++ packages/util/.npmignore
@@ -1,0 +1,3 @@
+/*
+!/dist/**/*.js
+*.test.js
