+++ packages/util/README.md
@@ -1,0 +1,5 @@
+# @xstyled/util
+
+xstyled core utilities. Kind of specific lodash for xstyled.
+
+See [xstyled documentation](https://www.smooth-code.com/open-source/xstyled/)
