+++ packages/util/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@xstyled/util",
+  "description": "xstyled utilities",
+  "version": "1.9.1",
+  "sideEffects": false,
+  "main": "dist/xstyled-util.cjs.js",
+  "module": "dist/xstyled-util.es.js",
+  "author": "Greg Bergé <berge.greg@gmail.com>",
+  "license": "MIT",
+  "scripts": {
+    "build": "rollup -c",
+    "prebuild": "shx rm -rf dist",
+    "prepublishOnly": "yarn build"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.4.4"
+  }
+}
