+++ packages/util/rollup.config.js
@@ -1,0 +1,8 @@
+import { getRollupConfig } from '../../config/rollup'
+import pkg from './package.json'
+
+export default getRollupConfig({
+  pwd: __dirname,
+  buildName: 'xstyled-util',
+  pkg,
+})
